Use localization for native and custom elements
